pyproject.toml
[project]
name = "uikit-agent"
version = "0.1.0"
description = "UIKit Agent - AI-powered code generation agent"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "huey[redis]",
    "redis",
    "python-dotenv",
    # LangChain ecosystem
    "langchain==1.0.8",
    "langchain-core==1.0.7",
    "langchain-openai==1.0.3",
    "langchain-anthropic>=1.1.0",
    "langchain-google-genai>=3.1.0",
    "langsmith==0.4.43",
    # LangGraph
    "langgraph==1.0.3",
    "langgraph-checkpoint==3.0.1",
    "langgraph-prebuilt==1.0.4",
    "langgraph-sdk==0.2.9",
    "uvloop>=0.22.1",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "rich>=14.2.0",
    "fastapi-limiter>=0.1.6",
    "mcp>=1.23.1",
    "anyio>=4.12.0",
    "sentry-sdk[fastapi]>=2.47.0",
]


[dependency-groups]
dev = [
    "ruff>=0.14.6",
]

[tool.ruff]
line-length = 150

[tool.ruff.lint]
# E, F - standard errors (pyflakes, pycodestyle)
# I - import sorting (isort)
# UP - pyupgrade (automatic syntax updates)
# B - flake8-bugbear - finds many hidden bugs
select = ["E", "F", "I", "B", "UP"]

# Import sorting configuration
[tool.ruff.lint.isort]
known-first-party = ["agents", "tasks", "schemas" ]


[tool.ruff.lint.flake8-bugbear]
# Add a list of functions that can be called in arguments
extend-immutable-calls = [
    "fastapi.Depends",
    "fastapi.Query",
    "fastapi.Path",
    "fastapi.Body",
    "fastapi.Security",
    "fastapi.Header",
]
